Day 6 - Tuning Trouble
Both parts turned out to be easier than I first expected when I started
reading the puzzle. Used a copied sub-list of the sequence and set() to
check for uniqueness. Only 1 char difference between the parts once
tidied up.

Keeping track of start and end indices and just passing a slice to set()
instead of keeping an actual frame aaround would be a nice improvement.