Day 11 - Monkey in the Middle
Enjoyed part 1 today. Knew immediately on reading part 2 that there was
going to be some clever modulo trick to keeping the numbers manageable.
Took me a long while to try a common multiple of all the divisors. Also,
I assume the primality of the divisors was important in some way?