Day 15 - Beacon Exclusion Zone
Pleased with my performance for part 1. Simplified to just checking
each sensor's distance to the target line and filling in the remaining
distance.

Part 2 was a nightmare. I wrote the most naive approach just to see
how slow it was. Then my first attempt at improving it was faster, but
still not feasible. Finally re-wrote using ranges, more similar to my
part 1 solution. Runs in 18s :(