Day 17 - Pyroclastic Flow
Tetris rocks! Very nice puzzle - think I did fairly well with part 1
once I'd stopped confusing myself with different coordiante systems
for the cave and the rocks. The naïve approach for part 2 was an
obvious non-starter (tried it anyway for fun). Eventually figured out
that as the rocks and jets repeat, the rate of height increase
becomes constant as soon as a repeated state (rock, jet & 'floor') is
found. Sadly, I wasn't handling the mean case Eric put in for rocks
slipping through which wasted a LOT of time.